fix: linking flags in the linux build system (#499)
* Replace arch by uname -m.

`uname -m` is a portable version of `arch`, which may not be
installed (see
https://www.gnu.org/software/coreutils/manual/coreutils.html#arch-invocation).

* Fix rustdeps linking in linux.

As can be seen in the Makefile, when building the rust dependencies on
Linux, a `--target` argument is used. This changes the target
directory, adding a _target triple_, as explained in
https://doc.rust-lang.org/cargo/guide/build-cache.html.
The preamble comments have been adjusted to fit this.

`rbspy` had some additionaly libraries missing in the preamble line,
this has also been fixed, and now both `pyspy` and `rbspy` have the
same loader flags.

Finally, the `Dockerfile` has also been adjusted to this new
format. In this case, since we cannot infer the target triple, the
library is just copied to both supported targets.

* Simplify the case matrix by moving the target library.

If the library is copied to the generic destination, the rest of the
build system can still be generic, no matter what the actual target
build directory is.