refactor(config): decouple internal config from public CLI config in exec and server/storage (#508)
* Decouple storage config from CLI server config.

Storage is currently using the server CLI configuration, which
presents some problems:
- Public configuration is coupled to internal configuration, making
evolving both configurations independently harder. This goes both
ways, e.g.:
  - Deprecated configuration variables are still accessible
  internally.
  - Adding internal configuration variables without exposing them
  publicly is hard.
 This problem is not specific of the storage component, so these
 changes could also be applied to other components (server, exec, etc.)
- Storage has access to a lot of unneeded server configuration. This
doesn't happen with other components.

The first problem already arised when the dbmanager was
implemented, but it may become worse with adhoc mode (where we may
want to use in-memory storage, for example).
This refactor tries to solve this problem by isolating public and
private configurations through a public API to build the internal
config (through NewConfig and With* methods, as needed).

This is just a PoC to validate (or reject) the approach, if it's
validated it should be extended to include other components, testing
should be added, etc.

* Add storage configuration test-suite.

* Decouple exec CLI configuration from the internal configuration.

Some of the defaults are now moved to the internal configuration
constructor (sample rate, logging level) and the type of exec mode is
made explicit (with exec as default). This last part will make it
easier to add new modes (e.g. adhoc mode).

* Follow the import style in use.