Fix the stack overflow on Alpine generated binaries.
Currently, running the published pyroscope binary or running a docker
container on exec mode to run Python programs causes a segmentation
fault, due to a stack overflow.
This doesn't seem to affect to binaries generated by other Linux
distributions, so Alpine setting a smaller default stack size may be
the cause.
The default is overriden with a 2MB stack, which seems to be enough to
solve the problem.