chore: add docker images for the developer guide in major Linux distributions (#532)
* Add development images for ubuntu latest.

Ubuntu latest is currently Ubuntu 20.04 (Focal). It ships go 1.13 by
default which causes some problems (no embed support) so go 1.16 is
installed instead. It also ships nodejs 10, which currently is below
the minimal required version, but it seem to work fine with it
anyway.

* Add development images for debian latest.

Debian stable (bullseye) doesn't ship with go1.16, but it includes it
in bullseye-backports packages.

* Add development images for centos latest.

Centos 8 includes nodejs 10, so it's currently failing.

* Add libunwind8-dev dependency to fix spy libraries linking.

* Add development images for fedora latest.

* Move the images to their own dir and explain why we do use them.