Refactor adhoc implementation to better support ingestion modes.
We generalize the different adhoc modes through a runner interface,
which allows to run all the different modes independently and also
integrate well with the existing code for exec/connect modes.

We also separate the file saving part, which should be then reused.

The configuration has been extended to support all the different
modes, but not all of them are yet supported.