feat: initial support for adhoc mode (#504)
* Add an example script that can be used to showcase adhoc mode.

The basic idea is to have a simple script that runs for some minutes
and gives some differences in the profiling data on different runs.
Eventually we should have examples in different languages too.

* Move the context creation in the server start to the caller.

Passing the context instead of creating it allows the caller to
control when to cancel the context to indicate that the server shoudl
be stopped. This will be useful when exec and server modes are
combined in adhoc mode.

* Combine exec and server modes into adhoc mode.

A new command adhoc is created that combines exec and server modes
into a single process. The approach is quite hacky, but it works:
- the server and the agent are run in a single process.
- once the agent finishes, the server is still running and allows to
visualize the data in the normal UI.

On the other hand, a lot of things are still missing:
- Only exec mode targets are supported.
- The agent uses the Upstream mode instead of Direct one.
- The configuration is also a combination of the exec and server
configurations, which is unnecessarily complex and presents some
problems. Depending the profiling data wants to be presented, many of
these configuration options will be unnecesary.
- The file-system is used to store the data, instead of using memory.

* Combine exec and server config options better.

Use squash to avoid prefixes, following the dbmanager configuration
approach.

* Move log level parsing to the config constructor.

We need to parse the log level once this way, making it slightly
simpler.

* Add support for an in-memory badger in the storage.

* Add an adhoc mode to exec config.

* Polish adhoc mode.

Several new features are implemented:
- The agent uses direct upstream instead of remote upstream.
- The server doesn't upstream its own data.
- In memory badger is used for the profiling data.
- The unneede CLI options are removed.

From the user point of view, these features give a reasonable approach
to the first step of adhoc mode, from the execution and ingestion
point of view, but persisting the generated data is still pending.

The solution has some shortcomings codewise. In order to support the
particularities of the adhoc mode, some dependency construction has
been moved outside of the components that use them. This is good, but
there's still quite a lot of coupling between different parts, and as
a result the way this extractions work now has room for improvement.
This include:
- The logger: it's constructed and then passed as a dependency to the
storage, server, and exec components. This is an attempt to unify how
logging is handled, but it's still not completely consistent.
- The storage: it's also constructed and then passed as a dependency
to the server and exec components. The rest of the exec modes don't
need a storage, but since there's no constructor for the exec, the way
to handle this is currently hacky. It may be better to actually split
the whole exec component by mode, but there would be also some
duplication that is not easy to isolate.

* Rename storage to st to avoid conflicts with package names.

This should appraise the linter.

* Fix Cli arguments in exec testsuite.

* Add an adhoc test case for exec config.

* Add tests for in-memory storage.

* Add basic support for JSON export.

* Add support to export in pprof and collapsed formats.

* Decouple server and record subcommands in adhoc mode.

Additionally, the configuration is now adhoc specific, which should
make it simpler. The server command doesn't do much now, as it only
contains an empty storage.

* Add initial support for push ingestion mode in adhoc.

This is a hacky implementation to experiment with the approach, it
needs to be improved when/if we validate the idea.

* Put adhoc subcommands under a feature flag.

* Use a more radical way to but adhoc behind a feature flag.

Instead of adding some flags that expose the subcommand anyway, a
environment variable is used, which completely hides it from the
casual user.

* Add a long description of how the adhoc subcommand should work.

This is an attempt to give a relatively easy way to integrate and
document all the possible ways in which profiling can work.

* Remove manual start for remote upstream.

This option makes the remote upstream inconsistent with the direct
upstream and it also harder to understand when it needs to be started
and stopped. It's already used incorrectly.

* Decouple exec, connect and adhoc.

* Fix typo.

* Refactor adhoc implementation to better support ingestion modes.

We generalize the different adhoc modes through a runner interface,
which allows to run all the different modes independently and also
integrate well with the existing code for exec/connect modes.

We also separate the file saving part, which should be then reused.

The configuration has been extended to support all the different
modes, but not all of them are yet supported.

* Revert all the changes made to the server.

We won't have a separate server for adhoc.

* Do not shadow package names.

* Disallow arguments in connect subcommand.

* Remove duplicated call to upstream start/stop

* Remove no longer existing disabled flags.

* Fix adhoc command handling.

* Make adhoc logging more human-friendly.

* Fix default spy name in adhoc exec mode.

* Fix imports.

* Abstract process waiting to a single function and support timeouts.

This is reused in exec, connect and also adhoc push.
Timeout support is also added, as a basis for a duration based
profiling.

* Fix connect testsuite.

* Refactor the ingestion handler to make it reusable from adhoc mode.

* Revert "Abstract process waiting to a single function and support timeouts."

It doesn't work exactly the same for push and exec and it pull will
also be slightly different, so the abstraction gets more convoluted
than going with just some copying.

* Move the initialization part to the constructor.

* Log the error runner error if any.

* Add initial support for adhoc pull mode.

* Fix package shadowing.

* Fix typo.