feat: Add adhoc push mode support in clib.
For native applications to work, they need to support the overriding
of the upstream address that ahdoc uses for this mode.

Instead of reusing PYROSCOPE_SERVER_ADDRESS enviroment variable,
a new variable PYROSCOPE_ADHOC_SERVER_ADDRESS has been used.
The users shouldn't be bothered with this override, using a separate
name should make it harder to see some undesired interactions.

Note that once this is merged the native integrations should be
upgraded to make use of this new clib for adhoc push to work.