feat: improve pyroscope data format.
Current data format used for rendering is not defined formally and has
a couple of issues:
- There's no format versioning, which makes it hard to evolve the
format or migrate between versions.
- Current format has redudancy, with duplicated metadata.

This may become a problem especially once adhoc mode is released, as
this format is directly exported to JSON.

I have created a new package with the flamebearer profile versioned
model, which should make the format more explicit and easier to
evolve. I have also removed the duplicated metadata.