feat(frontend): persist sidebar collapsed state (#699)
The logic is similar to what twitch does:

    If the user has never interacted with the sidebar (state called 'pristine'):
        If the window is small, we set to collapsed
        If the window width is large, we set to uncollapsed
        If the user resizes the screen, these rules should be still be valid.
    If the user HAS interacted with the sidebar
        then we assume they know what they are doing and ALWAYS use what've chosen
        for example, if they had manually collapsed, we will always have it collapsed, even in bigger screens

Also added migrations, which currently don't do much but with new versions this is useful. The idea is that what you got from localStorage may not be what your application code is expecting (let's say if you had stored a boolean but now your app code expects an object).

Create a ui store that currently only holds the sidebar state. Also that's the only state serialized.
Preferred a more straightforward approach to make the code easier to understand, the component interacts directly with the store via the selector and the dispatcher.