feat: output standalone HTML files for adhoc profiles. (#728)
* feat: output standalone HTML files for adhoc profiles.

* chore: improve flamebearer to standalone html function name.

* chore: fix typo in flamebearer

Thanks @eh-am for pointing out!

* feat: modify the way data is exported.

Instead of choosing between JSON, pprof and collapsed on one hand and
standalone HTML in the other, leave JSON on one hand and the rest in
the other. This makes more sense for two reasons:
- The main reason to generate pprof or collapsed formats is to either
interoperate with other tooling or to share the results. In this sense
it's more like an external format that can be used outside of
Pyroscope, like standalone HTML.
- It makes sense to always generate internal JSON data for better
results when the Pyroscope UI is used for analysis, even if an
external format is preferred for interoperability, i.e. JSON is always
a good companion to the other formats.

Thanks to @eh-am for the rationale.

* feat: remove the experimental status of adhoc.

It's still a feature that can be disabled and thus the UI doesn't
include Adhoc sections, but it's no longer experimental and it's
enabled by default.