refactor: make converted functions versioned. (#765)
This is a follow-up to #757. After some further discussion with @eh-am
we agreed that making these functions versioned gives a safer path to
handle version changes.

This is technically a backwards incompatible change that could be
avoided by creating both versioned and unversioned variants pointing
to the same code, but I think it's quite safe to break compatibility
here and avoid that burden.