fix: store "total" in name cache, and make tests more rigurous.
This is a minor fix, it just avoids having the name twice in the worst
case.

Also, name order is checked in the flamebearer tests. While the order
of the names is not important, the correspondence between name index
and name is, which means that without proper order checks it's not
possible to evaluate if the expectations are correct.