refactor: change the way snapshot callbacks handle errors. (#865)
* refactor: change the way snapshot callbacks handle errors.

Currently snapshot callbacks receive any error happening during data
gathering in the snapshot and they are in charge of handling them.

This inverts the usual responsibility chain where the caller should be
responsible to handle the errors of the called code, and instead
errors are passed between two unrelated code blocks.

This decision has two consequences:
- It violates the principle of least surprise. Code behaves
unexpectedly, which is arguably bad by itself.
- It reduces the flexibility in error handling. Making the callback
responsible of error handling is good enough to handle errors that
happen before the callback is called or during the callback call, but
it cannot handle the errors that happen after the callback finishes
but before the control is returned back to the caller, and this is
actually happening in our codebase (e.g. #849).

This commit changes the error handling to follow the usual flow of
making the caller responsible of handling the callee errors, and
allowing the callbacks to return those errors instead of handling
them.

This fixes #849.

* fix: don't check if the process exists with negatid PIDs.

A negative PID is used in ebpf to indicate whole system profiling.