feat(ingestion): initial support for JFR format ingestion. (#954)
* feat(ingestion): initial support for JFR format ingestion.

Java Flight Recorder format is a binary format integrated into Java
and used by AsyncProfiler, which is the native profiler used for our
Java integration.

It's the only AsyncProfiler format that supports profiling multiple
events concurrently and, in order to support this, we have created a
JFR parser.

This is a first iteration towards full, multi-event JFR ingestion
support. In this initial work, only CPU events (cpu, itimer, wall) are
supported. Support for other events should come later.

* remove debug output.

* chore(ingestion): move jfr parser to its own package.

This should prevent logrus to become a dependency of the go profiler.