fix(flamegraph): inject its styles via css only (#1023)
We were using style-loader, which calls insertStyleElement. That obviously doesn't work in a nodejs environment (not a react SSR env, but just plain node, which we use to allow people to render the flamegraph using node-canvas). We moved to css-loader, so now the only way to import css is by importin dist/style.css (which in node you simply would not do).