fix(frontend): fix frontend routing (#645)
Backend serves /ui/assets requests (css/js) using the embed file server.
All other routes are served the index.html, which then are handled via the frontend router (react-router).
When building for production, a BASEPATH env var is dynamically set in webpack to /ui. This is so that React Router knows that it needs to take that into account. This part will likely change when #638 is addressed.