feat: select first app automatically (#721)
When the application first starts, it loads the list of apps, and if available, sets the first one as the current query. This is so that users see something "useful" right away.

If user deletes the query later, it won't do anything (to not be annoying).

# Limitations
It doesn't update history in place, so if one hits the back button, it goes back to an empty query.