Add micro-service integration test
To test the integration of the microservices in a similar to is actually
run in the scalable/high availability setup.

After the cluster is fully started, it pushes profiles for a number of
services and then queries the series, label names and label values. It
then stops one instance each to ensure it is still queryable wihtout
data loss.