feat: Embed a grafana instance into pyroscope
This provides an optional target `embedded-grafana` to
download/configure/run a dedicated Grafana instance with Explore
Profiles deployed.

This is meant not to be used as production, but useful to quickly
discover profiles.

After this PR merges a user can run this:

```
$ docker run -p 4040:4040 -p 4041:4041 grafana/pyroscope --target all,embedded-grafana
```

And on localhost:4040 there will the classic pyroscope and on :4041,
there will be Grafana with Explore Profiles.