adhocprofile: Validate IDs and santize names properly
We currently allow any character to be user specified at upload. This
limits that character set. It also makes sure only IDs are listend and
retrieve with get, when they are fulfilling the character limit.

This validation will ensure that the generated identifiers are withing
the allowed set use for object store.