Fix: Handle context correctly
selectTree was using the a errgroup context going foward. This caused
some errors when store-gateways were used for queries, that required
deduplication.

This fix also updated all errGroup nested contexts to use gCtx, as an
additional measure we could consider using a block to limit the use
`gCtx` after `Wait()`.