fix(docs): Make sure helm Grafana example works as expected.
The problem with the grafana helm is that we don't have the profiling
port exposed as part of the containerPorts of the container. So I had to
reuse/abuse the gossip-tcp port in order to get it to be scraped.

It also was using the wrong annotation of "pyroscope.grafana.com"
instead of "profiles.grafana.com".

Fixes #3523