perf: cache stack hash to node in stacktreeTree
See the inline comment for more context.

Profiling against my workload showed the compactor spending significant
CPU time in `stacktraceTree.insert()`.

Staring at the code, I realized it was performing a full stack walk to
populate/resolve the leaf-most tree node representing the stack.

Since many stacks will be identical, we can benefit from caching
the mapping of the unique stack back to the node index.

The performance speedups are greater if there are many repeated stacks
and when stacks are deep. My data set consists of a lot of JVM profiles
and Java is infamous for having very deep stacks. So this yields a
considerable speedup for me.