Set GRAFANA_VERSION in front matter (#3971)
Currently all links with `<GRAFANA_VERSION>` in them are broken because that resolves to the current version of the Pyroscope documentation.
For example, `latest` in `/docs/pyroscope/latest/` but `v1.13.x` in `/docs/pyroscope/v1.13.x/`.

This uses `latest` as a reasonable default.
If specific versions of Pyroscope target specific versions of Grafana, it may be worth updating this on those branches to match.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>
(cherry picked from commit badbcfb834af42f2e1e8b3dc3b67174928b82a17)