Add benchmark for `*stacktraceTree.insert` with different init sizes
This benchmark can be slow, thus it won't run unless the
`COMPARE_STACKTRACETREE_INSERT_DEFAULT_SIZES` environment variable is
set to true. In addition, if the `-v` flag is passed to `go test` it
will print the initial and max tree size after inserting all the
samples from the test profile.