Compare `*stacktraceTree.insert` against different initial sizes (#4033)
* Use default size constant for initializing tree in benchmark

Refactor the benchmark for `*stacktraceTree.insert` to use the
constant for default tree size instead of hard-coding 0. This will
become useful in future changes.

* Add benchmark for `*stacktraceTree.insert` with different init sizes

This benchmark can be slow, thus it won't run unless the
`COMPARE_STACKTRACETREE_INSERT_DEFAULT_SIZES` environment variable is
set to true. In addition, if the `-v` flag is passed to `go test` it
will print the initial and max tree size after inserting all the
samples from the test profile.

* Remove commented out line

* Always run stacktraceTree default size benchmark

* Use b.ReportMetric instead of b.Log

We could revisit if we want to always show these values, and if we
want to add/remove columns.