chore: Update prometheus to version v3 (#4094)
* chore: Update prometheus to version v3.

Note this will relax the label validation and allow the newer metric selectors
format to allow utf8 symbols.

* Migrate testutil.NewTestingLogger

* Do usage group comparision by values

Instead fully equality of the matcher object, which breaks with the
latest prometheus version.

* No metric names are incorrect other than empty and non utf-8

* Use correct retry config

This change is a consequence by the grpc-go change in v1.70.0, which got
merged here: https://github.com/grpc/grpc-go/pull/7905

As we do call `WithDisableRetry()`, it should be a no-op.

* Remove method config