Improve `*deduplicatingSlice.ingest` performance (#4037)
* Add benchmark for `*PartitionWriter.WriteProfileSymbols`

I had to perform the initialization within the for loop as otherwise
it always panics.

* Refactor `*PartitionWriter.WriteProfileSymbols` for DRY

Thank you @simonswine for the suggestion!

Co-authored-by: Christian Simon <simon@swine.de>

* Create rewriting map with maximum capacity

By doing this we avoid having to regrow the map, which improves CPU
and memory consumption.

* Use temp slice to append only once

Instead of checking `s.slice` on each call to append and regrow if
necessary, create a temporary slice with the maximum possible
capacity, and append that slice to `s.slice` outside the for loop,
effectively checking and regrowing `s.slice` only once.

This further improves CPU and memory consumption.

* Use stdlib slices.Grow to grow the slice

---------

Co-authored-by: Christian Simon <simon@swine.de>