chore: Clean up collection rules store
This was added to store Alloy config related to Profile collection.
These settings were only stored there for a very brief moment and no
main commit or released version of the Profiles Drilldown app every
relied on it.

This clears up everything and will leave no trace. This is okay as the
feature was marked as experimental