fix: Breaking change in ProfileTypes v2
When updating the storage layer to v2, we introduced a breaking change
in the ProfileTypes API, while in v1 we would return, the last hour of
profile types, we stopped doing that in v2 and showed an error
instead:

```
{"code":"invalid_argument","message":"missing time range in the query"}
```

In Grafana Tempo's datasource configuration this call is used without
start/end parameters and therefore fails, see https://github.com/grafana/grafana/issues/110211.

I am proposing we maintain the old behaviour for ProfileTypes endpoint,
until we announce a breaking change change as part of a major release.

We also should ensure that Grafana uses it with start and end
parameters.