feat: Allow utf-8 characters in label names
This feature is a relaxation in previous validation for label names.
Given the backward incompatible nature of this change (for example,
PromQL queries must be [updated to support utf-8 label names](https://prometheus.io/docs/guides/utf8/#querying)),
it is gated behind a "client capability": a key/val that a client
must set in the `Accept:` header.