feat: Implement utf8 label name client capability (#4442)
This change implements a client capability framework, with `allow-utf8-labelnames` as the first capability. Capabilities are flags clients set in the `Accept` header to inform the API about specific client support. This is useful for cross-API features (like utf-8 support).

This change does _not_ provide full utf-8 label name support. Instead, it offers the first pass at what will be full support, with follow up changes to provide full support.

This change respects the `allow-utf8-labelnames` client capability in the read path (both v1 and v2 `LabelNames` and `Series` APIs). If the capability is not set or not set to true, these APIs will filter out "invalid" label names (i.e. with chars outside of [a-zA-Z0-9_.]) from the response. This logic is currently a no-op, given it's not yet possible to write label names outside of this charset.