fix: tryFindGo doesn't trim repo path when absolute path is used
When an absolute path is used like
`/home/christian/github.com/grafana/pyroscope/[....]`, we are not
matching the repository correctly and have to try quite a few files
before we truncated enough folders.

This change ensures that these cases are dealt with similar like a path
starting with the repo name.