fix: SelectSeries not correct for single reports (#4617)
* fix: SelectSeries not correct for single reports

The query-backend will only call the aggregate functions, when there is
more than one report for particular type. In the time series case the
aggregate function is also the one that ensures order/correctness of the
shape of data.

I considered doing this inside every query, but eventually it felt better to
do this when we would normally "aggregate"

* Ensure we are holding the lock when initializing the aggreagator.

* Use error