feat: Implement `java-jar-mapper` tool
This tool is used to modify a `.pyroscope.yaml` file with source
code mappings for 3rd party libraries from a given jar.

It uses heuristics and 3rd party APIs to try to resolve discovered
3rd party library functions in to git repos.

There is also a hard coded mapping in `jar-mappings.json` for any
popular 3rd party libraries that the tool's heuristics are unable
to resolve.